pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "djangocms-seo"
version = "0.1.0"
dependencies = [
    "django>=5.0.8",
    "django-cms>=4.1.2",
    "django-classy-tags>=4.1.0",
    "django-entangled>=0.5.4",
    "django-filer>=3.2.0",
    "django-formtools>=2.5.1",
    "django-fsm>=2.8.2",
    "django-parler>=2.3",
    "django-polymorphic>=3.1.0",
    "django-sekizai>=4.1.0",
    "django-treebeard>=4.7.1",
    "djangocms-admin-style>=3.3.1",
    "djangocms-alias>=2.0.1",
    "djangocms-attributes-field>=3.0.0",
    "djangocms-frontend>=1.3.3",
    "djangocms-text-ckeditor>=5.1.5",
    "djangocms-versioning>=2.0.2",
    "easy-thumbnails>=2.9",
    "tzdata>=2024.1",
]
description = "A Django CMS extension to manage SEO meta tags for web pages."
readme = "README.rst"
requires-python = ">=3.10"
authors = [
    { name = "Yasyn Ben Rhouma", email = "ben.rhouma.yessine0610@gmail.com" },
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: Django CMS",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]